Refactor helper methods. Move C# code out of YAML
