+++ resharper/resharper-unity/src/CSharp/Daemon/Stages/Analysis/LayerMaskAnalyzer.cs
@@ -8,7 +8,6 @@ using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches;
 using JetBrains.ReSharper.Plugins.Yaml.Settings;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.ReSharper.Psi.Tree;
-using static JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches.UnityProjectSettingsUtils;
 
 namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
 {
@@ -31,25 +30,22 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
 
         protected override void Analyze(IInvocationExpression element, ElementProblemAnalyzerData data, IHighlightingConsumer consumer)
         {
-            if (!myAssetSerializationMode.IsForceText) 
+            if (!myAssetSerializationMode.IsForceText)
                 return;
-            
+
             if (!myUnityYamlSupport.IsParsingEnabled.Value)
                 return;
-            
-            if (IsLayerMaskGetMask(element) || IsLayerMaskNameToLayer(element))
+
+            if (element.IsLayerMaskGetMaskMethod() || element.IsLayerMaskNameToLayerMethod())
             {
                 foreach (var argument in element.ArgumentList.Arguments)
                 {
-                    if (argument == null)
-                        return;
-
-                    var literal = (argument.Value as ICSharpLiteralExpression)?.ConstantValue.Value as string;
+                    var literal = (argument?.Value as ICSharpLiteralExpression)?.ConstantValue.Value as string;
                     if (literal == null)
                         return;
-                
+
                     var cache = element.GetSolution().TryGetComponent<UnityProjectSettingsCache>();
-                    if (cache != null && !cache.HasLayer(literal)) 
+                    if (cache != null && !cache.HasLayer(literal))
                         consumer.AddHighlighting(new UnknownLayerWarning(argument));
                 }
             }
