+++ resharper/resharper-unity/src/CSharp/Daemon/Stages/Analysis/LoadSceneAnalyzer.cs
@@ -8,7 +8,6 @@ using JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches;
 using JetBrains.ReSharper.Plugins.Yaml.Settings;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
 using JetBrains.ReSharper.Psi.Tree;
-using static JetBrains.ReSharper.Plugins.Unity.Yaml.Psi.Caches.UnityProjectSettingsUtils;
 
 namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
 {
@@ -38,23 +37,21 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
         {
             if (!myAssetSerializationMode.IsForceText)
                 return;
-            
+
             if (!myUnityYamlSupport.IsParsingEnabled.Value)
                 return;
-            
-            
-            var argument = GetSceneNameArgument(invocationExpression);
 
+            var argument = GetSceneNameArgument(invocationExpression);
             var literal = argument?.Value as ICSharpLiteralExpression;
             if (literal == null)
                 return;
-            
-            if (IsSceneManagerSceneRelatedMethod(invocationExpression.InvocationExpressionReference))
+
+            if (invocationExpression.InvocationExpressionReference.IsSceneManagerSceneRelatedMethod())
             {
                 var cache = invocationExpression.GetSolution().TryGetComponent<UnityProjectSettingsCache>();
                 if (cache == null)
                     return;
-                
+
                 var sceneName = GetScenePathFromArgument(literal);
                 if (sceneName != null)
                 {
@@ -70,8 +67,7 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
                         {
                             consumer.AddHighlighting(
                                 new LoadSceneAmbiguousSceneNameWarning(argument, sceneName));
-                        } 
-                        
+                        }
                     }
                     else
                     {
@@ -83,9 +79,8 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
                             consumer.AddHighlighting(new LoadSceneUnexistingSceneNameWarning(argument));
                         }
                     }
-
-                    
-                } else if (literal.ConstantValue.IsInteger())
+                }
+                else if (literal.ConstantValue.IsInteger())
                 {
                     var value = (int) literal.ConstantValue.Value;
                     if (value >= cache.SceneCount)
@@ -95,22 +90,45 @@ namespace JetBrains.ReSharper.Plugins.Unity.CSharp.Daemon.Stages.Analysis
                 }
             }
 
-
-            if (IsEditorSceneManagerSceneRelatedMethod(invocationExpression.InvocationExpressionReference))
+            if (invocationExpression.InvocationExpressionReference.IsEditorSceneManagerSceneRelatedMethod())
             {
                 var cache = invocationExpression.GetSolution().TryGetComponent<UnityProjectSettingsCache>();
                 if (cache == null)
                     return;
-                
+
                 var sceneName = GetScenePathFromArgument(literal);
                 if (sceneName != null)
                 {
                     if (!cache.IsSceneExists(sceneName))
                     {
                         consumer.AddHighlighting(new LoadSceneUnexistingSceneNameWarning(argument));
-                    } 
+                    }
                 }
             }
+        }
+
+        private static ICSharpArgument GetSceneNameArgument(IInvocationExpression invocationExpression)
+        {
+            return invocationExpression.ArgumentList.Arguments.FirstOrDefault
+                (t => t.IsNamedArgument && t.ArgumentName?.Equals("sceneName") == true || !t.IsNamedArgument);
+        }
+
+        private static string GetScenePathFromArgument(ICSharpLiteralExpression literalExpression)
+        {
+            // There are 3 ways to present scene name in unity
+            // Consider scene : Assets/Scenes/myScene.unity
+            // User could use "myScene", "Scenes/myScene" and "Assets/Scenes/myScene.unity" to load scene
+            // Internally, we work only with first and second format (see UnityProjectSettingsCache)
+
+            var constantValue = literalExpression.ConstantValue.Value as string;
+            if (constantValue == null)
+                return null;
+
+            var sceneName = constantValue;
+            if (sceneName.StartsWith("Assets/") && sceneName.EndsWith(UnityYamlFileExtensions.SceneFileExtensionWithDot))
+                sceneName = UnityProjectSettingsUtils.GetUnityScenePathRepresentation(sceneName);
+
+            return sceneName;
         }
     }
 }
